Restructure tables
