+++ client/src/components/tables/FileInfoTable.tsx
@@ -14,13 +14,13 @@ import EditOutlinedIcon from '@material-ui/icons/EditOutlined';
 import moment from 'moment';
 import { useSnackbar } from 'notistack';
 import React, { useEffect, useState } from 'react';
-import { File } from '../gql/documents';
-import { useDeleteFileMutation } from '../gql/queries';
-import { ActionType } from '../utils/constants';
-import { renderDuration, renderSource } from '../utils/enumRender';
-import { DeleteConfirmDialog } from './dialogs/DeleteConfirmDialog';
-import { FileForm } from './dialogs/FileForm';
-import { GenericError, NetworkError } from './ErrorSnackbars';
+import { File } from '../../gql/documents';
+import { useDeleteFileMutation } from '../../gql/queries';
+import { ActionType } from '../../utils/constants';
+import { renderDuration, renderSource } from '../../utils/enumRender';
+import { DeleteConfirmDialog } from '../dialogs/DeleteConfirmDialog';
+import { FileForm } from '../dialogs/FileForm';
+import { GenericError, NetworkError } from '../ErrorSnackbars';
 
 type Props = {
   file: File;
