+++ client/src/components/tables/SeriesRelatedDisplay.tsx
