+++ client/src/pages/EpisodePage/index.tsx
@@ -1,8 +1,8 @@
 import { Grid } from '@material-ui/core';
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { EpisodeInfo } from '../../components/EpisodeInfo';
-import { FileList } from '../../components/FileList';
+import { EpisodeInfoTable } from '../../components/tables/EpisodeInfoTable';
+import { FileInfoList } from '../../components/tables/FileInfoList';
 import { withAuth } from '../../utils/withAuth';
 
 const EpisodePage = () => {
@@ -11,10 +11,10 @@ const EpisodePage = () => {
     <div>
       <Grid container spacing={3}>
         <Grid item xs={12} key={`episode-page-1`}>
-          <EpisodeInfo episodeId={episodeId} key={`${episodeId}-info`} />
+          <EpisodeInfoTable episodeId={episodeId} key={`${episodeId}-info`} />
         </Grid>
         <Grid item xs={12} key={`episode-page-2`}>
-          <FileList episodeId={episodeId} key={`${episodeId}-files`} />
+          <FileInfoList episodeId={episodeId} key={`${episodeId}-files`} />
         </Grid>
       </Grid>
     </div>
