+++ client/src/pages/SeriesPage/index.tsx
@@ -1,9 +1,9 @@
 import { Grid } from '@material-ui/core';
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { SeriesInfo } from '../../components/SeriesInfo';
 import { EpisodesTable } from '../../components/tables/EpisodesTable';
-import { UserProgressInfo } from '../../components/UserProgressInfo';
+import { SeriesInfoTable } from '../../components/tables/SeriesInfoTable';
+import { UserProgressInfoTable } from '../../components/tables/UserProgressInfoTable';
 import { withAuth } from '../../utils/withAuth';
 
 const SeriesPage = () => {
@@ -13,10 +13,13 @@ const SeriesPage = () => {
     <div>
       <Grid container spacing={3}>
         <Grid item xs={12}>
-          <SeriesInfo seriesId={seriesId} key={`${seriesId}-info`} />
+          <SeriesInfoTable seriesId={seriesId} key={`${seriesId}-info`} />
         </Grid>
         <Grid item xs={12}>
-          <UserProgressInfo seriesId={seriesId} key={`${seriesId}-progress`} />
+          <UserProgressInfoTable
+            seriesId={seriesId}
+            key={`${seriesId}-progress`}
+          />
         </Grid>
         <Grid item xs={12}>
           <EpisodesTable seriesId={seriesId} key={`${seriesId}-table`} />
